refactor(acp-bridge): create skeleton + lift zero-coupling primitives (#4175 PR 22a)
First slice of #4175 Wave 5 PR 22 (`refactor(serve): extract acp bridge
primitives`). Lifts the three primitives from `packages/cli/src/serve/`
that have zero coupling to the rest of `serve/`, so PR 22a can land
ahead of PR 17 (#4282) and PR 14b (#4271) without merge-conflict risk
on `httpAcpBridge.ts`. Also seeds the `PermissionMediator` interface
contract that PR 24 will implement (4 strategies — first-responder /
designated / consensus / local-only — replacing the inline
first-responder voting in `BridgeClient.requestPermission`).

What moves
- `cli/src/serve/eventBus.ts` (578 LOC, no internal imports) →
  `packages/acp-bridge/src/eventBus.ts`
- `cli/src/serve/inMemoryChannel.ts` (73 LOC, only depends on
  `@agentclientprotocol/sdk`) → `packages/acp-bridge/src/inMemoryChannel.ts`
- `httpAcpBridge.ts:638-677` `AcpChannel` / `AcpChannelExitInfo` /
  `ChannelFactory` types → `packages/acp-bridge/src/channel.ts`
- Both test files moved alongside their sources

What's new
- `packages/acp-bridge/` package (`@qwen-code/acp-bridge`, internal,
  not published to npm yet)
- `packages/acp-bridge/src/permission.ts` — type-only
  `PermissionMediator` / `PermissionPolicy` / `PermissionResolution`
  / `PermissionRequestRecord` / `PermissionVote` /
  `PermissionVoteOutcome`. No implementation; PR 24 fills it in.
- `packages/cli/src/serve/eventBus.ts` and `inMemoryChannel.ts` are
  now one-line re-export wrappers for backward compat.
- `httpAcpBridge.ts:638-677` is now a one-line `import type` +
  `export type` re-export.

Backward compatibility
- All existing relative imports (`./eventBus.js`,
  `./inMemoryChannel.js`, `../serve/eventBus.js`) keep resolving via
  the wrappers — no churn for the 8 importer sites in
  `cli/src/serve/` plus `cli/src/commands/serve.ts:14`.
- `httpAcpBridge.ts` continues to export `AcpChannel` /
  `AcpChannelExitInfo` / `ChannelFactory` so any external consumer
  is unaffected.
- Zero `/capabilities` payload changes, zero HTTP route changes,
  zero SDK behavior changes, zero spawn-site behavior changes.

What's not in this slice (deferred to PR 22b after PR 17/14b land)
- `BridgeClient` + `createHttpAcpBridge` factory + `defaultSpawnChannelFactory`
  (~3000 LOC, lines 1082-3770 + 4106-4307 of `httpAcpBridge.ts`).
- Channel and VSCode-IDE-companion own-spawn migrations.

Tests
- `npm test --workspace @qwen-code/acp-bridge` — 28/28 pass
  (eventBus 20 + inMemoryChannel 8) at the new location.
- `cd packages/cli && npx vitest run src/serve/` — 567/567 pass
  (no regressions).
- `cd packages/cli && npx tsc --noEmit` clean.

References
- #4175 Wave 5 PR 22 row
- #3803 chiga0's "Stage 1.5-prereq AcpChannel lift"
- `httpAcpBridge.ts:679-696` FIXME for the four `PermissionMediator`
  strategies this slice declares
