+++ packages/acp-bridge/src/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright 2025 Qwen Team
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './eventBus.js';
+export * from './inMemoryChannel.js';
+export * from './channel.js';
+export * from './permission.js';
